230830 commit Http Response 예외처리 진행 1) GET /board/view?id={id} (글 번호 id의 내용 확인) 2) POST /board (글 등록) 3) PUT /board (글 수정) 4) DELETE /board?id={id}&password={password} (글 삭제)

+++ src/main/java/com/sideproject/makeboard/controller/BoardController.java
@@ -5,6 +5,8 @@ import com.sideproject.makeboard.dto.BoardInfoWithId;
 import com.sideproject.makeboard.dto.BoardInsertInfo;
 import com.sideproject.makeboard.dto.BoardUpdateInfo;
 import com.sideproject.makeboard.service.BoardService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 
 
 import lombok.extern.slf4j.Slf4j;
@@ -21,6 +23,8 @@ public class BoardController {
     private BoardController(BoardService boardService) {
         this.boardService = boardService;
     }
+
+    // 게시판 목록 전체 불러오기
     @GetMapping("list")
     public @ResponseBody List<BoardInfo> getBoardInfo(
             @RequestParam(value = "page", required = false)Long page,
@@ -31,24 +35,80 @@ public class BoardController {
         return boardService.getBoardInfo(page, listCnt);
     }
 
+    // 특정 게시판 불러오기
     @GetMapping("view")
-    public @ResponseBody BoardInfoWithId getBoardInfoWithId(@RequestParam(value = "id")Long id) {
+    public ResponseEntity<BoardInfoWithId> getBoardInfoWithId(@RequestParam(value = "id")Long id) {
+
         BoardInfoWithId boardInfoWithId = boardService.getBoardInfoWithId(id);
-        return boardInfoWithId;
+        if (!boardService.isExistsId(id)) {
+            return new ResponseEntity<BoardInfoWithId>(boardInfoWithId , HttpStatus.NOT_FOUND); // 404 Not Found
+        }
+        return new ResponseEntity<BoardInfoWithId>(boardInfoWithId , HttpStatus.OK);
+
     }
 
+    // 게시판에 글 생성
     @PostMapping
-    public void setBoardInfo(@RequestBody BoardInsertInfo boardInsertInfo) {
+    public ResponseEntity<String> setBoardInfo(@RequestBody BoardInsertInfo boardInsertInfo) {
+
+        if (boardInsertInfo.getNickname() == null || boardInsertInfo.getNickname().isEmpty()) {
+            return new ResponseEntity<String>("Nickname is required" , HttpStatus.BAD_REQUEST); //400 Code (ERROR)
+        }
+
+        if (boardInsertInfo.getPassword() == null || boardInsertInfo.getPassword().isEmpty()) {
+            return new ResponseEntity<String>("Password is required" , HttpStatus.BAD_REQUEST); //400 Code (ERROR)
+        }
+
+        if (boardInsertInfo.getTitle() == null || boardInsertInfo.getTitle().isEmpty()) {
+            return new ResponseEntity<String>("Title is required" , HttpStatus.BAD_REQUEST); //400 Code (ERROR)
+        }
+
         boardService.setBoardInfo(boardInsertInfo);
+        return new ResponseEntity<String>("New post is registered", HttpStatus.CREATED);
     }
 
+    // 게시판 내의 글을 수정
     @PutMapping
-    public void putBoardInfo(@RequestBody BoardUpdateInfo boardUpdateInfo) {
+    public ResponseEntity<String> putBoardInfo(@RequestBody BoardUpdateInfo boardUpdateInfo) {
+
+        if (boardUpdateInfo.getId() == null) {
+            return new ResponseEntity<String>("Id is required" , HttpStatus.BAD_REQUEST); //400 Code (ERROR)
+        }
+
+        if (boardUpdateInfo.getPassword() == null || boardUpdateInfo.getPassword().isEmpty()) {
+            return new ResponseEntity<String>("Password is required" , HttpStatus.BAD_REQUEST); //400 Code (ERROR)
+        }
+
+        if (boardUpdateInfo.getTitle() == null || boardUpdateInfo.getTitle().isEmpty()) {
+            return new ResponseEntity<String>("Title is required" , HttpStatus.BAD_REQUEST); //400 Code (ERROR)
+        }
+
+        if (!boardService.isExistsId(boardUpdateInfo.getId())) {
+            return new ResponseEntity<String>("Post Id : " +  boardUpdateInfo.getId().toString() + " does not exists", HttpStatus.NOT_FOUND); //404 Code (ERROR)
+        }
+
+        if (!boardService.isCorrectPw(boardUpdateInfo.getId(), boardUpdateInfo.getPassword())) {
+            return new ResponseEntity<String>("Password not correct", HttpStatus.BAD_REQUEST);
+        }
+
         boardService.putBoardInfo(boardUpdateInfo);
+        return new ResponseEntity<String>("Post is updated", HttpStatus.OK);
     }
 
+    // 게시판 내의 글을 삭제
     @DeleteMapping
-    public void deleteBoardInfo(@RequestParam(value = "id") long id, @RequestParam(value = "password")String password) {
+    public ResponseEntity<String> deleteBoardInfo(@RequestParam(value = "id") Long id,
+                                                  @RequestParam(value = "password")String password) {
+
+        if (!boardService.isExistsId(id)) {
+            return new ResponseEntity<String>("Post id : " + id.toString() + " does not exists" , HttpStatus.NOT_FOUND); // 404 Code (ERROR)
+        }
+
+        if (!boardService.isCorrectPw(id, password)) {
+            return new ResponseEntity<String>("Password not correct", HttpStatus.BAD_REQUEST); // 400 Code (ERROR)
+        }
+
         boardService.deleteBoardInfo(id, password);
+        return new ResponseEntity<String>("Post is deleted", HttpStatus.OK);
     }
 }
